Prevent flash-minting of tokens
Adding a check that asserts that within a transaction, we can not have a situation where tokens are
first added to an account, and only later subtracted from another.

If that were possible, you could insert a third accountupdate in the middle that would see the
increased balance and act upon it -- and there would be no limit as to how much the balance could
be. That could be relevant for stake-based voting, for example.