Allow updating the verification key.
This PR does two things:

- it adds a method `canChangeVerificationKey` to the admin contract, to check permissions for
changing the verification key. This method is called from `updateVerificationKey` in the token
contract.

- The deploy arguments for the token contract now have a new field, `allowUpdates`. If that is set
to `true`, updates of the verification key will be allowed also during the current protocol version.

This allows updating the token contract, for example when there is a new version of o1js.